fix(sales): deals filter with number
