+++ frontend/plugins/sales_ui/src/modules/deals/boards/components/list/DealsRecordTable.tsx
@@ -4,6 +4,7 @@ import { DealsColumn } from '@/deals/boards/components/list/DealsColumn';
 import { DealsCommandBar } from '@/deals/boards/components/list/DealsListCommandBar';
 import { NoStagesWarning } from '@/deals/components/common/NoStagesWarning';
 import { useDeals } from '@/deals/cards/hooks/useDeals';
+import { getDealsQueryVariables } from '@/deals/utils/queryVariables';
 import { useSearchParams } from 'react-router-dom';
 import { useStages } from '@/deals/stage/hooks/useStages';
 
@@ -18,26 +19,8 @@ export const DealsRecordTable = () => {
     skip: !pipelineId,
   });
 
-  const ignoredKeys = ['boardId', 'pipelineId', 'salesItemId', 'tab'];
-
-  const queryVariables: Record<string, any> = {};
-
-  for (const [key, value] of searchParams.entries()) {
-    if (ignoredKeys.includes(key)) continue;
-
-    try {
-      const parsed = JSON.parse(value);
-      queryVariables[key] = parsed;
-    } catch {
-      queryVariables[key] = value;
-    }
-  }
-
   const archivedOnly = searchParams.get('archivedOnly') === 'true';
-
-  if (archivedOnly) {
-    queryVariables.noSkipArchive = true;
-  }
+  const queryVariables = getDealsQueryVariables(searchParams);
 
   const { deals, loading, handleFetchMore } = useDeals({
     skip: !pipelineId,
